refactor: configuration 분리
